Move unit test to 'actor.test.ts'
